Removed mockito based user tests
Will need to look at a better solution to test these methods.
